fixed delete.php

+++ delete.php
@@ -15,24 +15,24 @@ $result = mysqli_query($conn,$sql);
 if($result->num_rows > 0){
 	$sql2 = "SELECT * from files where id ='$id';";
 	$result2 = mysqli_query($conn,$sql2);
-	if($row2 = $result2->fetch_assoc()){
 
-		$filetodelete = $row2['file_path'];
-		unlink($filetodelete);
-		$sql3 = "DELETE from files where id ='$id';";
-		$result3 = mysqli_query($conn,$sql3);
+	if($row2 = $result2->fetch_assoc()){
 		$archivedby = $row2['archived_by'];
 		$sql4 = "UPDATE files SET archive = '0' WHERE id= '$archivedby'";
-		//echo $sql4;
-                            if ($conn->query($sql4) === TRUE) {
-								header("location: admin/index.php?success=2");
-                            }else {
-                                echo "Error: " . $sql4 . "<br>" . $conn->error;
-							}	
+				if ($conn->query($sql4) === TRUE) {
+					$sql3 = "DELETE from files where id ='$id';";
+					if ($conn->query($sql3) === TRUE) {
+						$filetodelete = $row2['file_path'];
+						unlink($filetodelete);
+						header("location: admin/index.php?success=2");
+					}
+				}else {
+					echo "Error: " . $sql4 . "<br>" . $conn->error;
+				}	
+							
 	}else{
 
 	}
 }else{
 	header("location: admin/index.php?error=5");
 }
-?>
